feat : now can use string to create Component , but still have bugs

+++ Source/GameEngine/Game/Message/Source/Message.cpp
@@ -53,7 +53,7 @@ void Receiver::Clear()
 bool Receiver::IfHaveMessage(int c)
 {
 	bool re = false;
-	for (int i = 1; i <= m_MessageQueue.size(); i++)
+	for (size_t i = 1; i <= m_MessageQueue.size(); i++)
 	{
 		int buff = TakeOutMessage();
 		if (buff == c)
